+++ asst1/malloc.h
@@ -5,7 +5,7 @@
 #define malloc(x) mymalloc(x, __FILE__,__LINE__)
 #define free(x) myfree(x,__FILE__,__LINE__)
 
-void* mymalloc(int,char*,int);
+void* mymalloc(size_t,char*,int);
 void myfree(void*,char*,int);
 
 #endif
